TE - alembic updates

+++ alembic/versions/1f2e28f386b1_delete_map_image_assoc_table.py
@@ -7,6 +7,7 @@ Create Date: 2020-07-22 13:26:03.224366
 """
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy.engine.reflection import Inspector
 
 
 # revision identifiers, used by Alembic.
@@ -17,7 +18,11 @@ depends_on = None
 
 
 def upgrade():
-    op.drop_table('map_image_assoc')
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
+    if 'map_image_assoc' in tables:
+        op.drop_table('map_image_assoc')
 
 
 def downgrade():
